refactor: frontend skeleton
