+++ frontend/src/App.tsx
@@ -1,12 +1,10 @@
-import "./App.css"
-
-import ViewRoleListing from "./pages/applicant/ViewRoleListing"
+import ViewRoleListing from "./pages/applicant/main"
 
 function App() {
   return (
-    <>
+    <div>
       <ViewRoleListing />
-    </>
+    </div>
   )
 }
